refactor(gui): Stop every menu lookup casting for itself
Several places reached into items with the signed active_item and each would
have needed its own conversion to say that a row number is not negative. An
active_row() accessor makes that crossing once, next to the has_active_item()
test that is what makes it true, so the places that read a row now say which
row they mean rather than how to find it.

The rest are indexes that were already held to their container by the code
around them: the dialog's buttons, the string select's list, and the four
option lists, whose selectors are set either to a loop count or to the size of
a list about to be pushed to.

+++ src/gui/item_stringselect.cpp
@@ -100,7 +100,7 @@ ItemStringSelect::draw(DrawingContext& context, Vector pos, int menu_width, floa
                        LAYER_GUI);
   /* Centred rather than pushed against one arrow, so a short value does not
      sit at one end of the column with a hole beside it. */
-  context.draw_text(Resources::normal_font, list[*selected],
+  context.draw_text(Resources::normal_font, list[static_cast<size_t>(*selected)],
                     Vector(pos.x + centre, pos.y - int(Resources::normal_font->get_height()/2)),
                     ALIGN_CENTER, LAYER_GUI, active ? ColorScheme::Menu::active_color : get_color());
 }
